# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZeroQuant Trading API",
    "description": "\n# ZeroQuant 트레이딩 봇 REST API\n\n전략 관리, 백테스트, 포트폴리오 분석을 위한 REST API입니다.\n\n## 주요 기능\n\n- **전략 관리**: 트레이딩 전략 생성, 조회, 시작/중지\n- **백테스트**: 과거 데이터 기반 전략 성과 분석\n- **포트폴리오**: 실시간 포트폴리오 상태 조회\n- **시장 데이터**: 실시간 시세 및 차트 데이터\n- **모니터링**: 에러 추적 및 시스템 상태 모니터링\n- **스크리닝**: 종목 스크리닝 및 필터링\n- **ML**: 머신러닝 모델 훈련 및 배포\n\n## 인증\n\n대부분의 엔드포인트는 JWT Bearer 토큰 인증이 필요합니다.\n`Authorization: Bearer <token>` 헤더를 포함하세요.\n\n## 심볼 동기화\n\n- **KRX**: `POST /api/v1/dataset/sync/krx` - 한국 거래소 심볼\n- **EODData**: `POST /api/v1/dataset/sync/eod` - 해외 거래소 심볼\n",
    "contact": {
      "name": "ZeroQuant Team",
      "url": "https://github.com/user/zeroquant"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.4.4"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "로컬 개발 서버"
    }
  ],
  "paths": {
...
    "/api/v1/signals/by-strategy": {
      "get": {
        "tags": [
          "signals"
        ],
        "summary": "전략별 신호 조회",
        "operationId": "get_signals_by_strategy",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "query",
            "description": "전략 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_time",
            "in": "query",
            "description": "시작 시각 (ISO 8601)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "end_time",
            "in": "query",
            "description": "종료 시각 (ISO 8601)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "최대 결과 개수",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalSearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/signals/by-symbol": {
      "get": {
        "tags": [
          "signals"
        ],
        "summary": "심볼별 신호 조회",
        "operationId": "get_signals_by_symbol",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "description": "심볼 (예: \"005930\")",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exchange",
            "in": "query",
            "description": "거래소 (예: \"KRX\")",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_time",
            "in": "query",
            "description": "시작 시각 (ISO 8601)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "end_time",
            "in": "query",
            "description": "종료 시각 (ISO 8601)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "최대 결과 개수",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalSearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/signals/search": {
      "post": {
        "tags": [
          "signals"
        ],
        "summary": "지표 기반 신호 검색",
        "description": "JSONB 쿼리를 사용하여 특정 지표 조건을 만족하는 신호를 검색합니다.\n\n# 지원 연산자\n- `$gte`: >=\n- `$lte`: <=\n- `$gt`: >\n- `$lt`: <\n- `$eq`: =",
        "operationId": "search_signals",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignalSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "검색 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalSearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "description": "통합 API 에러 응답.\n\n모든 API 엔드포인트에서 일관된 에러 형식을 제공합니다.\n\n# 예시\n\n```json\n{\n  \"code\": \"STRATEGY_NOT_FOUND\",\n  \"message\": \"전략을 찾을 수 없습니다: rsi-123\",\n  \"details\": null,\n  \"timestamp\": 1738300800\n}\n```",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "에러 코드 (예: \"DB_ERROR\", \"INVALID_INPUT\", \"NOT_FOUND\")"
          },
          "details": {
            "description": "추가 에러 상세 정보 (선택적)"
          },
          "message": {
            "type": "string",
            "description": "사람이 읽을 수 있는 에러 메시지"
          },
          "method": {
            "type": [
              "string",
              "null"
            ],
            "description": "HTTP 메서드 (GET, POST 등)"
          },
          "path": {
            "type": [
              "string",
              "null"
            ],
            "description": "요청 경로"
          },
          "timestamp": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "에러 발생 타임스탬프 (Unix timestamp, 선택적)"
          }
        }
      },
...
      "RouteState": {
        "type": "string",
        "description": "종목의 매매 단계를 나타내는 5단계 상태.\n\n# 상태 설명\n\n- **Attack**: 진입 적기 - TTM Squeeze 해제 + 모멘텀 상승\n- **Armed**: 대기 준비 - Squeeze 중 + 강한 기본 구조\n- **Wait**: 관찰 중 - 정배열 유지, 저가 상승 추세\n- **Overheat**: 과열 - 급등 후 조정 필요\n- **Neutral**: 중립 - 명확한 신호 없음",
        "enum": [
          "ATTACK",
          "ARMED",
          "WAIT",
          "OVERHEAT",
          "NEUTRAL"
        ]
      },
...
      "SignalIndicators": {
        "type": "object",
        "description": "신호 생성에 사용된 기술적 지표 값들.",
        "properties": {
          "atr": {
            "type": [
              "string",
              "null"
            ],
            "description": "ATR (Average True Range)"
          },
          "bb_lower": {
            "type": [
              "string",
              "null"
            ],
            "description": "볼린저 밴드 하단"
          },
          "bb_middle": {
            "type": [
              "string",
              "null"
            ],
            "description": "볼린저 밴드 중간"
          },
          "bb_upper": {
            "type": [
              "string",
              "null"
            ],
            "description": "볼린저 밴드 상단"
          },
          "breakout_score": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "돌파 점수"
          },
          "ema_long": {
            "type": [
              "string",
              "null"
            ],
            "description": "EMA (장기)"
          },
          "ema_short": {
            "type": [
              "string",
              "null"
            ],
            "description": "EMA (단기)"
          },
          "macd": {
            "type": [
              "string",
              "null"
            ],
            "description": "MACD"
          },
          "macd_histogram": {
            "type": [
              "string",
              "null"
            ],
            "description": "MACD 히스토그램"
          },
          "macd_signal": {
            "type": [
              "string",
              "null"
            ],
            "description": "MACD 시그널"
          },
          "range_pos": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "박스권 내 위치 (0.0 ~ 1.0)"
          },
          "route_state": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RouteState",
                "description": "RouteState (매매 단계)"
              }
            ]
          },
          "rsi": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "RSI (14일)"
          },
          "sma_long": {
            "type": [
              "string",
              "null"
            ],
            "description": "SMA (장기)"
          },
          "sma_short": {
            "type": [
              "string",
              "null"
            ],
            "description": "SMA (단기)"
          },
          "squeeze_momentum": {
            "type": [
              "string",
              "null"
            ],
            "description": "Squeeze 모멘텀"
          },
          "squeeze_on": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Squeeze 상태 (압축 중)"
          },
          "vol_quality": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "거래량 품질"
          }
        }
      },
      "SignalMarkerDto": {
        "type": "object",
        "description": "신호 마커 응답 DTO",
        "required": [
          "id",
          "symbol",
          "timestamp",
          "signal_type",
          "price",
          "strength",
          "indicators",
          "reason",
          "strategy_id",
          "strategy_name",
          "executed"
        ],
        "properties": {
          "executed": {
            "type": "boolean",
            "description": "실행 여부"
          },
          "id": {
            "type": "string",
            "description": "신호 ID"
          },
          "indicators": {
            "$ref": "#/components/schemas/SignalIndicators",
            "description": "지표 정보"
          },
          "price": {
            "type": "string",
            "description": "가격"
          },
          "reason": {
            "type": "string",
            "description": "신호 이유"
          },
          "side": {
            "type": [
              "string",
              "null"
            ],
            "description": "방향 (Buy/Sell)"
          },
          "signal_type": {
            "type": "string",
            "description": "신호 유형"
          },
          "strategy_id": {
            "type": "string",
            "description": "전략 ID"
          },
          "strategy_name": {
            "type": "string",
            "description": "전략 이름"
          },
          "strength": {
            "type": "number",
            "format": "double",
            "description": "신호 강도 (0.0 ~ 1.0)"
          },
          "symbol": {
            "type": "string",
            "description": "심볼"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "타임스탬프"
          }
        }
      },
      "SignalSearchRequest": {
        "type": "object",
        "description": "지표 기반 검색 요청",
        "required": [
          "indicator_filter"
        ],
        "properties": {
          "indicator_filter": {
            "$ref": "#/components/schemas/Value",
            "description": "지표 필터 (JSONB 쿼리)\n\n# 예시\n```json\n{\n  \"rsi\": {\"$gte\": 70.0},\n  \"macd\": {\"$gt\": 0}\n}\n```"
          },
          "limit": {
            "type": "integer",
            "format": "int64",
            "description": "최대 결과 개수 (기본 100, 최대 1000)"
          },
          "signal_type": {
            "type": [
              "string",
              "null"
            ],
            "description": "신호 유형 필터 (선택)"
          }
        }
      },
      "SignalSearchResponse": {
        "type": "object",
        "description": "신호 검색 응답",
        "required": [
          "total",
          "signals"
        ],
        "properties": {
          "signals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SignalMarkerDto"
            },
            "description": "신호 목록"
          },
          "total": {
            "type": "integer",
            "description": "총 결과 수",
            "minimum": 0
          }
        }
      },
...
      "Value": {}
    }
  },
...
}